# docs/api_spec.yaml
openapi: 3.0.0
info:
  title: Distributed Synchronization System API
  version: 1.0.0
  description: Dokumentasi API untuk Lock Manager, Queue, dan Cache System.
servers:
  - url: http://localhost:8001
    description: Node 1
  - url: http://localhost:8002
    description: Node 2
  - url: http://localhost:8003
    description: Node 3
paths:
...
  /cache/write:
    post:
      summary: Menulis data ke cache (dan memori utama)
      tags: [CacheCoherence]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address: { type: string, example: "kunci_A" }
                value: { type: string, example: "data_rahasia" }
      responses:
        '200':
          description: Data berhasil ditulis
          content:
            application/json: { example: {"status": "written", "address": "kunci_A", "value": "data_rahasia"} }
  /cache/read:
    get:
      summary: Membaca data dari cache
      tags: [CacheCoherence]
      parameters:
        - name: address
          in: query
          required: true
          schema:
            type: string
            example: "kunci_A"
      responses:
        '200':
          description: Data ditemukan di cache (Hit)
          content:
            application/json: { example: {"status": "hit", "address": "kunci_A", "value": "data_rahasia"} }
        '404':
          description: Data tidak ditemukan (Miss)
          content:
            application/json: { example: {"status": "miss", "address": "kunci_A", "value": null} }